IoTFirmwareFuzzer/requirements.txt
# IoT Firmware Fuzzer - Python Dependencies
pyyaml>=6.0.1
psutil>=5.9.8
click>=8.1.7
colorlog>=6.8.2
rich>=13.7.0
numpy>=1.26.4
scipy>=1.12.0
scapy>=2.5.0
paho-mqtt>=2.0.0
capstone>=5.0.1
pyelftools>=0.31
jinja2>=3.1.3
flask>=3.0.0
aiofiles>=23.2.1
pytest>=8.0.0
pytest-asyncio>=0.23.5
